requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# HTTP and API
requests==2.31.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0

# Machine Learning and AI
torch==2.1.0
transformers==4.36.0
tokenizers==0.15.0
huggingface-hub==0.19.0
Pillow==10.1.0
numpy==1.24.3

# Image processing
opencv-python==4.8.1.78
scikit-image==0.22.0

# Configuration and environment
python-dotenv==1.0.0
configparser==6.0.0
PyYAML==6.0.1

# Logging and monitoring
loguru==0.7.2
psutil==5.9.6

# Utilities
click==8.1.7
tqdm==4.66.1
rich==13.7.0
typer==0.9.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For advanced features
redis==5.0.1
celery==5.3.4

